Move the `calculateSHA384` and `calculateSHA512` functions into their own file
This allows us to remove a closure, and we also change the code to initialize various constants lazily.
